debug/requirements.txt
 # =========================
# Core Backend
# =========================
fastapi==0.110.0
uvicorn[standard]~=0.41.0
pydantic~=2.12.5
python-dotenv==1.0.1

# =========================
# Database
# =========================
psycopg[binary]==3.1.18
SQLAlchemy==2.0.27
alembic==1.13.1

# =========================
# eBay API / HTTP
# =========================
httpx==0.26.0
requests==2.31.0

# =========================
# NLP (Query Parsing)
# =========================
spacy==3.7.2

# =========================
# Semantic Retrieval
# =========================
sentence-transformers==2.5.1
faiss-cpu==1.7.4
numpy==1.26.4
scikit-learn==1.4.0
google-genai==1.65.0

# =========================
# Trust Scoring / ML Light
# =========================
transformers==4.38.1
torch==2.2.1

# =========================
# Utils
# =========================
tqdm==4.66.2
pandas==2.2.0
rapidfuzz~=3.14.3
beautifulsoup4~=4.14.3



faiss-cpu

passlib~=1.7.3

jose~=1.0.0
mcp~=1.26.0
redis~=7.3.0


fastembed